- better looking queue tooltip - more appropriate drag pixmap
git-svn-id: svn://svn.dolezel.info/fatrat/trunk/fatrat@107 78086a28-8538-0410-bde1-fa2ac3882955

+++ engines/TorrentProgressWidget.cpp
@@ -45,7 +45,7 @@ QImage TorrentProgressWidget::generate(const std::vector<bool>& data, int width,
 		}
 		while(from <= to);
 		
-		int rcolor = 255 - qMin(color, 255.0);
+		quint32 rcolor = 255 - qMin(quint32(color), 255U);
 		buf[i] = 0xff0000ff | (rcolor << 8) | (rcolor << 16);
 	}
